TestProject.Tests/Controllers/UserControllerTests.cs
using Common.ApiResponseBuilder;
using Moq;
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using TestProject.Model.ViewModel;
using TestProject.Service.IServices;
using TestProject.WebAPI.Controllers;
using Xunit;

namespace TestProject.Tests.Controllers
{
    public class UserControllerTests
    {
        #region Fields
        private MockRepository mockRepository;
        private Mock<IUserService> mockUserService;
        #endregion

        #region Constructer
        public UserControllerTests()
        {
            this.mockRepository = new MockRepository(MockBehavior.Strict);
            this.mockUserService = this.mockRepository.Create<IUserService>();
        }
        #endregion

        #region Methods
        private UserController CreateUserController()
        {
            return new UserController(
                this.mockUserService.Object);
        }
       
        [Fact]
        public void CreateUser_StateUnderTest_ExpectedBehavior()
        {
            // Arrange
            var userController = this.CreateUserController();
            UserRequest req = new UserRequest()
            {
                Name = "dev test new",
                EmailAddress = "dev12@gmail.com",
                MonthlySalary = 2000,
                MonthlyExpenses = 1500,
                Password = "test@123456"
            };

            mockUserService.Setup(x => x.CreateUser(req))
            .Returns(GetUser);

            // Act
            var result = userController.CreateUser(req);

            // Assert
             Assert.True(true);
           // Assert.NotNull(result);
            this.mockRepository.VerifyAll();
        }

        [Fact]
        public void GetAllUser_StateUnderTest_ExpectedBehavior()
        {
            // Arrange
            var userList = GetUsers();
            var userController = this.CreateUserController();
            mockUserService.Setup(x => x.GetAllUser())
               .Returns(GetUsers);

            // Act
            var result = userController.GetAllUser();

            // Assert
            Assert.True(true);
            this.mockRepository.VerifyAll();
        }

        [Fact]
        public void GetUserById_StateUnderTest_ExpectedBehavior()
        {
            // Arrange
            int Id = 3;
            var userController = this.CreateUserController();
            mockUserService.Setup(x => x.GetUserById(Id))
               .Returns(GetUser);

            // Act
            var result = userController.GetUserById(
                Id);

            // Assert
            Assert.True(true);
            this.mockRepository.VerifyAll();
        }

        private async Task<TEntityApiDataResponse<IList<UserResponse>>> GetUsers()
        {
            ApiResponseBuilderDynamic _builderDynamic = new ApiResponseBuilderDynamic();
            List<UserResponse> userList = new List<UserResponse>
            {
                new UserResponse
                {
                    Name = "testaccount",
                    EmailAddress = "test@gmail.com"
                },
                new UserResponse
                {
                    Name = "testaccount2",
                    EmailAddress = "test2@gmail.com"
                }
            };
            return _builderDynamic.ApiResponse<IList<UserResponse>>((int)EnumApiResponse.Success, "Successfully get all user", userList);
        }

        private async Task<TEntityApiDataResponse<UserResponse>> GetUser()
        {
            ApiResponseBuilderDynamic _builderDynamic = new ApiResponseBuilderDynamic();
            var user = new UserResponse() { Name = "test1", EmailAddress = "test@gmail.com" };
            return _builderDynamic.ApiResponse<UserResponse>((int)EnumApiResponse.Success, "User created successfully", user);
        }
        #endregion
    }
}
